requirements.txt
numpy
scipy
pandas
python-dotenv
pyotp
smartapi-python